backend/requirements.txt
alembic==1.18.4
email-validator==2.3.0
fastapi==0.129.0
passlib[argon2]==1.7.4
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic-settings==2.13.0
python-dotenv==1.2.1
python-jose[cryptography]==3.5.0
python-multipart==0.0.22
requests==2.32.5
SQLAlchemy==2.0.46
uvicorn[standard]==0.41.0